Refactor: AdminChallengeList 템플릿 분리
