+++ src/pages/AdminChallengeList/template/style.ts
